login logout change
css videoplayer,
shifted google login to login route
navbar image of google
edited some shit in videoplayer.js
